+++ src/backend/cuda/transform/lower_l2_persistent_annotation.cc
@@ -1,5 +1,5 @@
 /*!
- * \file lower_l2_persistent_annotation.cc
+ * \file tl/backend/cuda/transform/lower_l2_persistent_annotation.cc
  * \brief Lower L2 persistent annotation
  */
 
@@ -9,8 +9,8 @@
 #include <tvm/tir/stmt_functor.h>
 #include <tvm/tir/transform.h>
 
-#include "../op/builtin.h"
-#include "../runtime/runtime.h"
+#include "backend/cuda/runtime.h"
+#include "op/builtin.h"
 
 namespace tvm {
 namespace tl {
