+++ tilelang/backend/cuda/op/gemm/__init__.py
@@ -1,10 +1,12 @@
+"""CUDA GEMM op registrations."""
+
 from __future__ import annotations
 
-from tilelang.backend.gemm import register_gemm_impl
-from tilelang.tileop.gemm.gemm_mma import GEMM_INST_MMA, GemmMMA
-from tilelang.tileop.gemm.gemm_mma_sm70 import GemmMMASm70
-from tilelang.tileop.gemm.gemm_tcgen05 import GEMM_INST_TCGEN05, GemmTCGEN5
-from tilelang.tileop.gemm.gemm_wgmma import GEMM_INST_WGMMA, GemmWGMMA
+from tilelang.tileop.gemm.registry import register_gemm_impl
+from .gemm_mma import GEMM_INST_MMA, GemmMMA
+from .gemm_mma_sm70 import GemmMMASm70
+from .gemm_tcgen05 import GEMM_INST_TCGEN05, GemmTCGEN5
+from .gemm_wgmma import GEMM_INST_WGMMA, GemmWGMMA
 from tilelang.utils.target import target_is_cuda, target_is_volta
