+++ tilelang/backend/cuda/op/gemm/gemm_mma.py
@@ -1,37 +1,44 @@
-from .gemm_sp_base import GemmSPBase
+from __future__ import annotations
+
+from tilelang.tileop.gemm.gemm_base import GemmBase
 from tilelang.layout import make_swizzled_layout
-from tilelang.intrinsics.mma_sp_macro_generator import SparseTensorCoreIntrinEmitter
-from tilelang.utils.language import is_shared, is_fragment
+from tilelang.intrinsics.mma_macro_generator import (
+    TensorCoreIntrinEmitter,
+)
+from tilelang.utils.language import is_shared, is_fragment, is_full_region
 from tilelang import tvm as tvm
 from tvm.target import Target
+from tvm.ir import Range
 from tvm import tir
 from tilelang import language as T
 from tilelang.transform.simplify import _Simplify
 
 
-GEMM_SP_INST_MMA = "cuda.mma"
+GEMM_INST_MMA = "cuda.mma"
 
 
-class GemmSPMMA(GemmSPBase):
-    def infer_layout(self, target: Target, thread_nums: int):
-        # NOTE(wt): Actually gemm_sp v2 currently use GemmWarpPolicy
-        m_warp, n_warp = self.policy.compute_warp_partition(self.M, self.N, thread_nums, target, GEMM_SP_INST_MMA)
+class GemmMMA(GemmBase):
+    def _make_mma_emitter(self, target: Target, thread_nums: int, thread_var: tir.Var | None = None):
+        m_warp, n_warp = self.policy.compute_warp_partition(self.M, self.N, thread_nums, target, GEMM_INST_MMA)
         warp_row_tiles = int(self.M // m_warp)
         warp_col_tiles = int(self.N // n_warp)
-        mma_emitter = SparseTensorCoreIntrinEmitter(
+        emitter = TensorCoreIntrinEmitter(
             a_dtype=self.in_dtype,
-            e_dtype=self.e_dtype,
             b_dtype=self.in_dtype,
             accum_dtype=self.accum_dtype,
             a_transposed=self.trans_A,
             b_transposed=self.trans_B,
-            e_transposed=self.trans_E,
             block_row_warps=m_warp,
             block_col_warps=n_warp,
             warp_row_tiles=warp_row_tiles,
             warp_col_tiles=warp_col_tiles,
-            warp_k=self.K,
+            chunk=self.chunk,
+            thread_var=thread_var,
         )
+        return emitter
+
+    def infer_layout(self, target: Target, thread_nums: int):
+        mma_emitter = self._make_mma_emitter(target, thread_nums)
         if self.is_gemm_ss():
             return {
                 self.A: make_swizzled_layout(self.A),
@@ -59,40 +66,40 @@ class GemmSPMMA(GemmSPBase):
         else:
             raise ValueError(f"Unsupported gemm combination, A: {self.A.scope()}, B: {self.B.scope()}")
 
-    def lower(self, target: Target, thread_nums: int, thread_var: tir.Var):
-        # NOTE(wt): Actually gemm_sp v2 currently use GemmWarpPolicy
-        m_warp, n_warp = self.policy.compute_warp_partition(self.M, self.N, thread_nums, target, GEMM_SP_INST_MMA)
-        warp_row_tiles = int(self.M // m_warp)
-        warp_col_tiles = int(self.N // n_warp)
-        mma_emitter = SparseTensorCoreIntrinEmitter(
-            a_dtype=self.in_dtype,
-            b_dtype=self.in_dtype,
-            e_dtype=self.e_dtype,
-            accum_dtype=self.accum_dtype,
-            a_transposed=self.trans_A,
-            b_transposed=self.trans_B,
-            e_transposed=self.trans_E,
-            block_row_warps=m_warp,
-            block_col_warps=n_warp,
-            warp_row_tiles=warp_row_tiles,
-            warp_col_tiles=warp_col_tiles,
-            warp_k=self.K,
-            thread_var=thread_var,
-        )
+    def lower(
+        self,
+        layout_map: dict,
+        target: Target,
+        thread_bounds: Range,
+        thread_var: tir.Var,
+        mbar_phase_expr: tir.PrimExpr | None = None,
+    ):
+        thread_nums = thread_bounds.extent
+        mma_emitter = self._make_mma_emitter(target, thread_nums, thread_var=thread_var)
 
         in_dtype = self.in_dtype
         warp_rows = mma_emitter.warp_rows
         warp_cols = mma_emitter.warp_cols
         local_size_a = mma_emitter.local_size_a
-        local_size_e = mma_emitter.local_size_e
         local_size_b = mma_emitter.local_size_b
+        block_K = mma_emitter.chunk
         micro_size_k = mma_emitter.micro_size_k
-        A_shared = self.ARegion
-        E_shared = self.ERegion
-        B_shared = self.BRegion
-        C_local = self.C
+        # We use region for memory input to support strided gemm
+        # T.gemm(A_shared[0:128, :], B_shared, C_local)
+        A_region = self.ARegion
+        B_region = self.BRegion
+        C_region = self.CRegion
+
+        A_buf = A_region.buffer
+        B_buf = B_region.buffer
+        C_buf = C_region.buffer
+
         clear_accum = self.clear_accum
-        assert micro_size_k <= self.K, f"K dimension {self.K} should be >= micro size k {micro_size_k}"
+
+        assert block_K >= micro_size_k, f"block_K ({block_K}) must be >= micro_size_k ({micro_size_k})"
+
+        assert is_full_region(C_region), "Fragment output C must be a full region"
+
         if self.is_gemm_ss():
 
             @T.prim_func
@@ -103,42 +110,32 @@ class GemmSPMMA(GemmSPBase):
                 accumulating into C_local.
                 """
                 A_local = T.alloc_local((warp_rows * local_size_a), in_dtype)
-                E_local = T.alloc_local((warp_rows * local_size_e), self.e_dtype)
                 B_local = T.alloc_local((warp_cols * local_size_b), in_dtype)
-
                 if clear_accum:
-                    T.clear(C_local)
-
-                for ki in T.serial(0, (self.K // micro_size_k)):
+                    T.clear(C_buf)
+                for ki in T.serial(0, (block_K // micro_size_k)):
                     # Load A into fragment
                     mma_emitter.ldmatrix_a(
                         A_local,
-                        A_shared,
-                        ki,
-                    )
-
-                    # Load E into fragment
-                    mma_emitter.ldmatrix_e(
-                        E_local,
-                        E_shared,
+                        A_region,
                         ki,
                     )
 
                     # Load B into fragment
                     mma_emitter.ldmatrix_b(
                         B_local,
-                        B_shared,
+                        B_region,
                         ki,
                     )
 
                     # Perform Matrix Multiplication
-                    mma_emitter.mma_sp(A_local, E_local, B_local, C_local, ki)
+                    mma_emitter.mma(A_local, B_local, C_buf, ki)
 
             # Simplify to optimize the index computing
             # Must inline let statements to simplify the analysis
             return _Simplify(_gemm_ssr, inline_let=True)
         elif self.is_gemm_sr():
-            B_local = self.B
+            assert is_full_region(B_region), "Fragment input B must be a full region"
 
             @T.prim_func
             def _gemm_srr() -> None:
@@ -148,28 +145,19 @@ class GemmSPMMA(GemmSPBase):
                 accumulating into C_local.
                 """
                 A_local = T.alloc_local((warp_rows * local_size_a), in_dtype)
-                E_local = T.alloc_local((warp_rows * local_size_e), self.e_dtype)
 
-                if clear_accum:
-                    T.clear(C_local)
-
-                for ki in T.serial(0, (self.K // micro_size_k)):
+                for ki in T.serial(0, (block_K // micro_size_k)):
+                    if clear_accum:
+                        T.clear(C_buf)
                     # Load A into fragment
                     mma_emitter.ldmatrix_a(
                         A_local,
-                        A_shared,
-                        ki,
-                    )
-
-                    # Load E into fragment
-                    mma_emitter.ldmatrix_e(
-                        E_local,
-                        E_shared,
+                        A_region,
                         ki,
                     )
 
                     # Perform Matrix Multiplication
-                    mma_emitter.mma_sp(A_local, E_local, B_local, C_local, ki)
+                    mma_emitter.mma(A_local, B_buf, C_buf, ki)
 
             # Simplify to optimize the index computing
             # Must inline let statements to simplify the analysis
@@ -177,7 +165,7 @@ class GemmSPMMA(GemmSPBase):
             # insert into parent block
             return _Simplify(_gemm_srr, inline_let=True)
         elif self.is_gemm_rs():
-            A_local = self.A
+            assert is_full_region(A_region), "Fragment input A must be a full region"
 
             @T.prim_func
             def _gemm_rsr() -> None:
@@ -186,36 +174,26 @@ class GemmSPMMA(GemmSPBase):
                 B_shared into local fragments, then issues Tensor Core mma ops,
                 accumulating into C_local.
                 """
-                E_local = T.alloc_local((warp_rows * local_size_e), self.e_dtype)
                 B_local = T.alloc_local((warp_cols * local_size_b), in_dtype)
-
                 if clear_accum:
-                    T.clear(C_local)
-
-                for ki in T.serial(0, (self.K // micro_size_k)):
-                    # Load E into fragment
-                    mma_emitter.ldmatrix_e(
-                        E_local,
-                        E_shared,
-                        ki,
-                    )
-
+                    T.clear(C_buf)
+                for ki in T.serial(0, (block_K // micro_size_k)):
                     # Load B into fragment
                     mma_emitter.ldmatrix_b(
                         B_local,
-                        B_shared,
+                        B_region,
                         ki,
                     )
 
                     # Perform Matrix Multiplication
-                    mma_emitter.mma_sp(A_local, E_local, B_local, C_local, ki)
+                    mma_emitter.mma(A_buf, B_local, C_buf, ki)
 
             # Simplify to optimize the index computing
             # Must inline let statements to simplify the analysis
             return _Simplify(_gemm_rsr, inline_let=True)
         elif self.is_gemm_rr():
-            A_local = self.A
-            B_local = self.B
+            assert is_full_region(A_region), "Fragment input A must be a full region"
+            assert is_full_region(B_region), "Fragment input B must be a full region"
 
             @T.prim_func
             def _gemm_rrr() -> None:
@@ -224,21 +202,10 @@ class GemmSPMMA(GemmSPBase):
                 B_shared into local fragments, then issues Tensor Core mma ops,
                 accumulating into C_local.
                 """
-                E_local = T.alloc_local((warp_rows * local_size_e), self.e_dtype)
 
-                if clear_accum:
-                    T.clear(C_local)
-
-                for ki in T.serial(0, (self.K // micro_size_k)):
-                    # Load E into fragment
-                    mma_emitter.ldmatrix_e(
-                        E_local,
-                        E_shared,
-                        ki,
-                    )
-
+                for ki in T.serial(0, (block_K // micro_size_k)):
                     # Perform Matrix Multiplication
-                    mma_emitter.mma_sp(A_local, E_local, B_local, C_local, ki)
+                    mma_emitter.mma(A_buf, B_buf, C_buf, ki)
 
             # Simplify to optimize the index computing
             # Must inline let statements to simplify the analysis
