+++ tilelang/backend/cuda/op/gemm_sp/__init__.py
@@ -1,7 +1,9 @@
+"""CUDA sparse GEMM op registrations."""
+
 from __future__ import annotations
 
-from tilelang.backend.gemm_sp import register_gemm_sp_impl
-from tilelang.tileop.gemm_sp.gemm_sp_mma import GemmSPMMA
+from tilelang.tileop.gemm_sp.registry import register_gemm_sp_impl
+from .gemm_sp_mma import GemmSPMMA
 from tilelang.utils.target import target_is_cuda
