+++ tilelang/backend/rocm/__init__.py
@@ -1,1 +1,1 @@
-from . import gemm  # noqa: F401
+from . import op  # noqa: F401
