+++ tilelang/backend/rocm/op/__init__.py
@@ -1,0 +1,1 @@
+from . import gemm  # noqa: F401
